more work on reply form
