+++ packages/client/src/Compose/RecipientsInput.tsx
@@ -1,22 +1,16 @@
 /* eslint-disable default-case */
 
-import { Chip, InputAdornment, makeStyles, TextField } from "@material-ui/core"
+import { InputAdornment, TextField } from "@material-ui/core"
 import { TextFieldProps } from "@material-ui/core/TextField"
 import { parseAddressList, ParsedGroup, ParsedMailbox } from "email-addresses"
 import * as React from "react"
-import Avatar from "../Avatar"
+import ParticipantChip from "../ParticipantChip"
 
 export type Address = ParsedMailbox
 
 type Props = TextFieldProps & {
   onRecipients: (recipients: Address[]) => void
 }
-
-const useStyles = makeStyles(theme => ({
-  chip: {
-    margin: theme.spacing(0.5, 0.25)
-  }
-}))
 
 type Action =
   | { type: "add"; recipient: Address }
@@ -68,7 +62,6 @@ function reducer(state: State, action: Action): State {
 }
 
 export default function RecipientsInput({ onRecipients, ...rest }: Props) {
-  const classes = useStyles()
   const [{ recipients, inputValue }, dispatch] = React.useReducer(reducer, {
     recipients: [],
     inputValue: ""
@@ -86,20 +79,9 @@ export default function RecipientsInput({ onRecipients, ...rest }: Props) {
         startAdornment: (
           <InputAdornment position="start">
             {recipients.map(recipient => (
-              <Chip
+              <ParticipantChip
                 key={email(recipient)}
-                avatar={
-                  <Avatar
-                    address={{
-                      name: recipient.name,
-                      mailbox: recipient.local,
-                      host: recipient.domain
-                    }}
-                  />
-                }
-                tabIndex={-1}
-                label={display(recipient)}
-                className={classes.chip}
+                recipient={recipient}
                 onDelete={() => dispatch({ type: "remove", recipient })}
               />
             ))}
@@ -124,10 +106,6 @@ export default function RecipientsInput({ onRecipients, ...rest }: Props) {
       value={inputValue}
     />
   )
-}
-
-function display({ name, address }: Address): string {
-  return name ? `${name} <${address}>` : address
 }
 
 function email({ address }: Address): string {
