+++ packages/client/src/documents/operations.graphql
@@ -55,7 +55,7 @@ mutation authenticate($id: ID!) {
   }
 }
 
-query getConversation($id: ID!) {
+query getConversation($id: ID!, $accountId: ID!) {
   conversation(id: $id) {
     id
     presentableElements {
@@ -88,6 +88,23 @@ query getConversation($id: ID!) {
     }
     isRead
     labels
+    replyRecipients(fromAccountId: $accountId) {
+      from {
+        name
+        mailbox
+        host
+      }
+      to {
+        name
+        mailbox
+        host
+      }
+      cc {
+        name
+        mailbox
+        host
+      }
+    }
     snippet
     subject
   }
